refactor rwc_nst_test.h; TxPower now works, as does LBT.db
